🚚 refactor: refactor with better position
