+++ src/app/chat/(mobile)/features/SessionList.tsx
@@ -1,15 +1,9 @@
 import { memo } from 'react';
 
-import { useSessionStore } from '@/store/session';
-
 import SessionListContent from '../../features/SessionListContent';
 import SessionSearchBar from '../../features/SessionSearchBar';
 
 const Sessions = memo(() => {
-  const useFetchSessions = useSessionStore((s) => s.useFetchSessions);
-
-  useFetchSessions();
-
   return (
     <>
       <div style={{ padding: '8px 16px' }}>
