+++ src/app/chat/components/SidebarHeader/index.tsx
