+++ src/app/chat/page.tsx
@@ -1,8 +1,8 @@
-import SessionHydration from '@/components/StoreHydration/SessionHydration';
 import { isMobileDevice } from '@/utils/responsive';
 
 import DesktopPage from './(desktop)';
 import MobilePage from './(mobile)';
+import SessionHydration from './components/SessionHydration';
 import Migration from './features/Migration';
 
 const Page = () => {
